requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.11.9
python-multipart==0.0.6

# External services
weaviate-client==4.16.10
openai==1.58.1
supabase==2.10.0

# HTTP client
httpx==0.26.0

# Environment and configuration
python-dotenv==1.0.0

